requirements.txt
python-telegram-bot
requests
faiss-cpu
sentence-transformers
pdfplumber